+++ xoeuf/fields/__init__.py
@@ -23,3 +23,9 @@ from .timezone import TimezoneSelection  # noqa: reexport
 from .timedelta import TimeDelta         # noqa: reexport
 from .one2one import One2one             # noqa: reexport
 from .timerange import TimeRange         # noqa: reexport
+
+
+try:
+    Serialized  # noqa
+except NameError:
+    from xoeuf.odoo.addons.base_sparse_field.models.fields import Serialized # noqa